Update buffer lengths to reflect actual hex output length, add debug assert code.
Some of the buffer lengths were longer than needed, which wastes SRAM when being called.  Fixed.

Added DEBUG_ASSERT compile time #define, for programmer use - check to ensure that the strings being written are not longer than the buffer (as that's all sorts of bad).  Add extra test code to print the longest strings to verify that this is working properly.